move get_loader_options function to test_integration_train.py
